+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2026-03-10
 
 - feat(cli): CLI entry point with vault init/list/status/unlink
+- feat(vault): add vault config command
 
 ## 2026-03-09